Crate smart_config_derive

Procedural macros for smart-config.

All macros in this crate are re-exported from the smart-config crate. See its docs for more details and the examples of usage.

Macro reference

Examples of usage

Derive Macros

DescribeConfig

Derives the DescribeConfig trait for a type. Typically used together with DeserializeConfig to generate a fully deserializable, self-descriptive configuration.

DeserializeConfig

Derives the DeserializeConfig trait for a type. Typically used together with DescribeConfig to generate a fully deserializable, self-descriptive configuration.

ExampleConfig

Derives the ExampleConfig trait for a type. This allows to provide an example for all config parameters.